requirements.txt
pandas==1.3.5
torch==1.12.0
torchvision==0.13.0
torchaudio==0.12.0
Pillow==9.0.1
scikit-learn==0.24.2
pytorch-lightning==2.0.2
torchmetrics==0.11.4
easyocr==1.6.2
opencv-python==4.5.5.64
opencv-contrib-python==4.6.0.66
opencv-python-headless==4.5.4.60
imutils==0.5.4
pytesseract==0.3.10
transformers==4.20.1